Use markdown for all instructions including Docker
This commit includes instructions for using the Docker container on
each platform (CentOS/Ubuntu/Mac/Windows). The instructions have been
updated to recommend an IP address that will not go away as external
network connectivity changes. This means that the HNN GUI will remain
running even after a laptop screen is closed or Wi-Fi locations have
changed.

The procedure for Windows is slightly simplified by using a consistent
IP address for the DISPLAY environment variable. Docker Desktop will
create this interface and it will likely be the same for all users.