Split install instructions into different files
In order to make it more apparent that the docker-based installation is
the recommended method, separate the native install instructions into a
separate file that a user would have to intentionally chose (marked for
advanced users).

Include instructions on creating a Docker Hub account in Mac and
Windows guides.

Also separate troubleshooting page for Mac into its own file.