Don't install HNN in /usr/local/
To run HNN, use the documented Docker procedure, or from the commandline:
python3 hnn.py

This is part of a move to not require sudo privileges to install HNN and
is relevant to the native install only. Significant modifications to
native install instructions, so Ubuntu and CentOS will probably need to
be tested more thoroughly.

Resolves #27