ENH: use physical cores by default
If hyperthreading is turned on, then multiprocessing.cpu_count()
will return 2x the number of actual cores. This causes a
performance slowdown with NEURON. So use psutil to detect the
number of physical cores. If the user specifies a different number
of cores, then keep that value.

Also if mpiexec fails with the given number of cores, try halving
and if successful, use that number for future simulations.