ENH: add range slider for opt ranges
This commits updates the configure optimization dialog with the most
notable change being the presence of a slider to select the range
over which the optimization algorithm will explore for that
parameter.

The way that the configure optimization dialog gets constructed has
been refactored only calling the relevant parts when "Recalculate
Ranges" is pressed.