BUG: fix improper use of str.join()
This was a mistake in using the join() method on a list. Now the
system also has adequeate time to check for killed process. The
fact that this error happened at all means there was a problem
stopping the processes. This is the best we can do within Python.

Fixes issue #109