ENH: add hnn_docker.sh for Mac Docker Desktop
Convert to a new method of starting HNN from within a docker
container for Mac. Instead of connecting to the X server directly,
use an ssh connection between the host (with X server) and the
container to form a reliable connection. This requires a lot of
setup commands to be run.

This scheme implements X11 authentication correctly, so the
option in XQuartz to disable authentication is no longer needed.
In fact, this script will set all XQuartz options automatically.

The hnn_docker.sh script removes the need for the user to run
docker and docker-compose commands. Additionally, it adds
options for upgrading HNN, stopping HNN and uninstalling HNN.

Partially fixes #125
Partially fixes #9