BUG: don't rely on default xhost in Linux
This fixes an edge case with username-based xauth as used in Linux
by switching to Xauthority file for X11 authorization on all
platforms. Refactor hnn_docker.sh to perform some of the same steps
for all OS's.

If an OS sets the environment variable and that file exists, use
it instead of forcing the use of ~/.Xauthority.

Related to #161