TST: large update to .travis.yml
Includes the following changes:

TST: mac enable testing back to El Capitan

Testing of docker is done through docker-machine. Since Travis
doesn't support VTX (needed for VirtualBox), we use qemu to
create VM that is not hardware-accelerated. A prebuilt qemu
binary for osx mojave is fetched, while qemu is build from
source for high sierra, sierra, and el capitan.

The docker-machine and qemu driver for docker-machine are included
in the fetched tarball with qemu. This avoids needing to install
go for building them.

The code for docker-machine has been updated to allow a longer time
waitin for the VM to start

Also the code for docker-machine-driver-qemu has been updated to allow
caching mode and use the -netdev syntax compatible with qemu 4.2.

TST: linux: test mpiexec in container
TST: linux: test all Ubuntu distros