BUG: reliably start sshd when using hnn_docker.sh
This commit improves hnn_docker.sh to check that sshd is running in
the container, and if not, it will start sshd and log messages from
sshd to /tmp/sshd.log within the container.

Fixes some output bugs in the script too: removing ^M characters and
a bad output redirect.

Dockerfile is updated to allow specifying SOURCE_REPO from the command
line for development builds using downstream forks. Also converting to
ADD syntax for neuron binary instead of using wget.

NEURON_VERSION can be specified as a build arg too.