ENH: QT file dialog window improvements
Add shortcuts to hnn_out and hnn_source_code to the GUI dialog that
opens when users select files to load for data or params.

Also limit the files to the appropriate type (.param or .txt) and
don't give a bogus warning when they click cancel from the load
parameter file dialog.