MAINT: cleanup external viz windows
For windows that are launched as a separate process, only read the param
file once with hnn-core and pass params as necessary. This avoids the
need for globals in most cases.

Globals still exist in visrast.py -- more work needed there

Also this completes using 'spec_cmap' with hnn core param files