MAINT: pass params dict to RunSimThread
When the user clicks the button to run a simulation, the parameters from the
GUI are immediately saved to a file using the __str__ function of each Qt
dialog box. Then using hnn-core.read_params(), the dictionary of parameters
is returned. This dictionary is passed to RunSimThread for both regular
simulations and optimization batches.

The QComboBox functions are simplified and updated for the new format of
simdat.lsimdat