MAINT: reimplement saving figures
Large change that refactors plotting functions in simdat.py
to reuse code when plotting in the GUI and when creating a
plot to save to a .png.

Only two types of figures are saved: dpl and spec. Other figures
can be added in the future. Another limitation is displaying
poisson inputs along with evprox and evdist; it's not supported
right now.

The spectrograms are configured to cut of the first 50ms to
avoid edge effects. This produces not so visually pleasing figures
when saving. In the GUI, the dipole also loses the first 50ms!
A better way of choosing the time range for the spectograms needs
to be devices.