Update installation instructions, add caveat
This updates the documented HNN-original installation methods for most
platforms to a newer version based on
https://github.com/jonescompneurolab/hnn/pull/337 . This is necessary
because:

1. The most recent install instructions found on the official HNN
website
https://hnn.brown.edu/installation-instructions/
provide only either a link to #337 or, in Oscar's case, a separate set
of instructions. However, these webpages on the official HNN website
will soon be deleted, since we will soon be doing a comprehensive
overhaul of the HNN websites. Once this happens, the only guidance that
users will receive about installation will need to be present on the
README of the repository here
https://github.com/jonescompneurolab/hnn

2. The Installation page indicated on the README of the repository
https://github.com/jonescompneurolab/hnn
does NOT include the install guidance from #337. This PR takes the old
pre-existing instructions for each platform, moves them into respective
subdirectories such as `2021_instructions` (the last year they were
updated), and then updates most of the instructions for each platform to
match #337.

I successfully tested the updated install versions for brown_ccv, mac,
ubuntu, and windows, and was able to get them all working using the
provided instructions. For Mac and Windows, I built off of
https://github.com/jonescompneurolab/hnn-tutorials/pull/4

This also adds a disclaimer at the top of every install page that
redirects users to HNN-Core. Even people in the lab have still been
getting confused about which install instructions are for which
version (original HNN or HNN-Core), so that should be cleared up now.