[MRG]: DOC: Final install instructions (#346)
* Relocate pre-existing install instructions

* Update installation instructions, add caveat

This updates the documented HNN-original installation methods for most
platforms to a newer version based on
https://github.com/jonescompneurolab/hnn/pull/337 . This is necessary
because:

1. The most recent install instructions found on the official HNN
website
https://hnn.brown.edu/installation-instructions/
provide only either a link to #337 or, in Oscar's case, a separate set
of instructions. However, these webpages on the official HNN website
will soon be deleted, since we will soon be doing a comprehensive
overhaul of the HNN websites. Once this happens, the only guidance that
users will receive about installation will need to be present on the
README of the repository here
https://github.com/jonescompneurolab/hnn

2. The Installation page indicated on the README of the repository
https://github.com/jonescompneurolab/hnn
does NOT include the install guidance from #337. This PR takes the old
pre-existing instructions for each platform, moves them into respective
subdirectories such as `2021_instructions` (the last year they were
updated), and then updates most of the instructions for each platform to
match #337.

I successfully tested the updated install versions for brown_ccv, mac,
ubuntu, and windows, and was able to get them all working using the
provided instructions. For Mac and Windows, I built off of
https://github.com/jonescompneurolab/hnn-tutorials/pull/4

This also adds a disclaimer at the top of every install page that
redirects users to HNN-Core. Even people in the lab have still been
getting confused about which install instructions are for which
version (original HNN or HNN-Core), so that should be cleared up now.

* docker: add ssh port fix

This applies Dylan's fix to today's issue with Mac usage of the
suggested installation instructions, as mentioned in this comment:
https://github.com/jonescompneurolab/hnn/pull/346#issuecomment-3330245902

The cause of the bug appears to be thus: in
`scripts/docker_functions.sh`, `get_container_port()` and elsewhere
depends on identifying the port of the HNN container for use. However,
the current combination of `cut` and a regex assume from the use of a
command like

```
docker port hnn_container 22
```

as used here:
https://github.com/jonescompneurolab/hnn/blob/e0cf45f876761280062511b56e7de088e7950b01/scripts/docker_functions.sh#L617

would output something like

```
0.0.0.0:49669
::49669
```

possibly excluding the second line, I'm not sure. The first line is an
IPv4 version of the container's local IP and its port, and the second is
an IPv6 version.

However, even though the documention doesn't make mention of the IPv6
output ( https://docs.docker.com/reference/cli/docker/container/port/ ),
it appears that as part of Docker Engine v28
https://docs.docker.com/engine/release-notes/28/#2800
the IPv6 output was changed to include brackets. This now affects
current use of `docker port`. For example, on my current up-to-date
version of Docker Desktop (using Docker Engine Community 28.4.0),
running `docker port hnn_container 22` gives me the following output:

```
0.0.0.0:49669
[::]:49669
```

The new brackets displayed, and/or the IPv6 output displayed, break the
bash scripting we use to identify the port. The new bash scripting in
this commit should fix this.

* doc: tiny windows instruction edition

* doc: fix ubuntu install issues